python-chatbot-backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
PyPDF2==3.0.0
pdf2image==1.16.3
pytesseract==0.3.10
docx2txt==0.8.0
transformers==4.35.0
torch==2.1.0
pinecone==7.3.0
python-dotenv==1.1.1
pymupdf==1.26.4
python-multipart==0.0.20
